requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
pdfkit==1.0.0
python-slugify==7.0.0
requests==2.28.1
sendwithus==5.3.0
six==1.16.0
text-unidecode==1.3
urllib3==1.26.14
